Create tests for SluggishWand
Added the three test methods: shouldApplyHeavyStatusToCreatureOnAttack,
to check that when a player attacks a target with the wand equipped a
HEAVY status will be applied to the target;
shouldReduceBaseSpeedOfCreatureOnAttack, to check that the targeted
creature based target's base speed has been halved when attacked by a
player with the wand equipped; shouldReduceBaseSpeedOfPlayerOnAttack, to
check that the targeted player's base speed has been halved when
attacked by a player with the wand equipped. Also added SluggishWand
stub class to allow the creation of the tests.