Add the flowerpot
The Flowerpot class got implemented by specifying the onUse method. Upon
usage it will create a VegetableGrower that grows carrots in its place.
Then the item was added into the game by creating and entry  in the
container.xml file and registering the entity in EntityMap.java and
EntityViewFactory.java as per stendhal add item guidelines.

Flowerpot.png was created to serve as a graphical representation for the
item