Add FastFerryTicket to the game
Added the ticket into the game that is required for the player to have
to get on the ferry. The information for the ticket has been added to
documents.xml and it has been registered in both EntityMap.java and
EntityViewFactory.java.

Additionally PortManagerNPCTest and FerryConveyerNPC have been updated,
allowing the player to be given a ticket based on how the
FastFerryTicket has been implemented.

Issue: #14