Implement base functionality of Wand of Sluggishness
The wand can now be summoned in game by admin commands, and will apply a
heavy status effect on the target that slows them down. This works for
both creatures and other players. Had to add new method to RPEntity to
deal with removing added StatusAttackers from the wand on unequip.
Refactored tests to deal with the true behaviour of the codebase in
regards to how it handles items, equipping and status effects.