Add ferry functionality and restructure tests
Added the interaction with the NPCs Eliza, in Ados coast, and Fiete, in
Deniran river. They can be spoken to (by saying "deniran" or "ados"
after saying "hi") and if the player has a ticket for the ferry they
will be teleported to the ferry destination.

The tests have also been restructured. The tests for checking if the
player has been teleported and if the NPC says they can get on the ferry
if they have a ticket have been merged into one test, as the teleport
directly follows the reply. Also the way the tests are set up have been
modified due to an issue with teleporting the player in the way the
tests were previously set up.

Issue: #14