Set maxRipness correctly
Upon closer examination of the code we have concluded that the resulting
vegetable was ripe, ready to grow, done. We reckoned that this would
lead to an unbalanced mechanic that could ruin the game's macro-economy
so we have altered the FlowerPot and the VegetableGrower to accept
maxRipeness as an attribute instead of having it set to 1. Also we
edited the VegetableGrower describe function to accommodate the correct
messages.