Fix tests and implement parameters and tests
There was an issue in some of our tests in that the original
implementations of actions were being used instead of the migrated one.
This has now been fixed and tests pass. We have also now accounted for
null values in XML actions which we didn't before, causing some tests to
fail.

Additionally tests for getMinimumParameters and getMaximumParameters
have been created and functionality for them have been implemented. In
XMLAction we have created attributes for min and max parameters and
these can be set through DefaultAction that is passed into the class.