Edit HashMap statements to use new Class XMLAction
This change was made so that the actions HashMap continues to store the
correct objects in it. Before it took in instances of the classes of
each action, however, now these classes will be removed and so this
change was needed to make the code work correctly. This seemed like the
easiest and simplest way to change it.