Move SlashActionRepository changes to a new file
The changes we made to SlashActionRepository have been moved to a new
class XMLSlashActionRepository. This is to ensure that current test all
pass and there is still functionality as the new system has not been
fully migrated to yet. Additionally some of the values for creating the
new classes have been updated based on the values used in the XMLs.